say command is still die but also new ban alias

+++ src/commands/owner/say.ts
@@ -18,6 +18,13 @@ export default class say extends Command {
                     unordered: true,
                     type: 'channel'
                 },
+                {
+                    id: "nodelete",
+                    match: 'option',
+                    flag: '--nodelete',
+                    unordered: true,
+                    type: 'bool'
+                }
             ],
             channel: 'guild'
         });
@@ -28,14 +35,15 @@ export default class say extends Command {
         if (args.channel) {
             args.channel.send(args.say)
             message.channel.send(`Message sent!`)
-            console.log(`Message ${args.say} was sent in ${message.channel.name} on ${message.guild.name} by ${message.author.tag}`)
+            console.log(`Message ${args.say} was sent in #${message.channel.name} on ${message.guild.name} by ${message.author.tag}`)
         }
+
         else {
             message.channel.send(args.say)
-            if (message.deletable) {
+            if (message.deletable && args.nodelete != true) {
                 await message.delete()
             }
-            console.log(`Message ${args.say} was sent in ${message.channel.name} on ${message.guild.name} by ${message.author.tag}`)
+            console.log(`Message ${args.say} was sent in #${message.channel.name} on ${message.guild.name} by ${message.author.tag}`)
         }
     }
 }
